+++ samples/node/object_factory.js
@@ -4,9 +4,12 @@ var class4js = require("../../lib/class4js.js");
 
 var Person = $class({
   
-  __construct__: function () {
+  __construct__: function (props) {
     this.__name = null;
     this.__age = 0;
+    if (props) {
+      $init(this, props);
+    }
   },
 
   name: {
@@ -29,9 +32,10 @@ var Person = $class({
 
 });
 
-var person1 = new Person();
-person1.name = "John Smith";
-person1.age = 30;
+var person1 = new Person({
+  name: "John Smith",
+  age: 30
+});
 console.log(person1);
 
 var person2 = $create(Person, {